+++ src/Services/Customer.API/Migrations/CustomerContextModelSnapshot.cs
@@ -23,11 +23,11 @@ namespace Customer.API.Migrations
 
             modelBuilder.Entity("Customer.API.Entities.Customer", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
                     b.Property<string>("EmailAddress")
                         .IsRequired()
